Made PStack & FStack subclasses of _Stack
- breaking change for the FStack cons method
  - before: my_stack.cons(None) == my_stack
  -    now: my_stack.cons(None) == None

In the spirit of this package: How does one add a non-existent value
to the beginning of a stack? In the case of immutable stacks, this is
similar to the case of taking the Cartesian product of two sets where
one of the sets is the empty set. This is the main reason I took cons
away from PStack.

What to work a litle on the annotations. The package code is coming up
clean but the LSP is coloring up the testcode red a bit more than
I would like.

Also may want to give more thought to the Map familyof methods.