WARNING: CLArray API is in flux
- finished overhauling __repr__ & __str__ methods
- finished __repr__ tests for all data structures

Originally saw CLArray (Constant Length Array) as a component of
a vector space, but now I think I will do something else for that,
maybe in a separate repository.

CLArray now matches the other data structures by not allowing
None values to be contained in it.

Having size > 0 is an unnecessary limitation. Also makes the data
structure a swmigroup and not a monoid after changing addition from
component-wise to concatenations. Also the default values to replace
empty slots (and now None values) give the data structure a wierd type
vibe. Not sure where to go with this. It is an interesting data
structure and I hope to find a use for it.